initializing registry users in the DB only once for all ITs

+++ registry-integration-tests/src/test/java/org/gbif/registry/database/DatabaseCleaner.java
@@ -137,9 +137,11 @@ public class DatabaseCleaner implements BeforeAllCallback {
               .execute();
         }
       }
-      connection
-          .prepareStatement("TRUNCATE " + String.join(",", tablesNotChallengeCode) + " CASCADE")
-          .execute();
+      if (!tablesNotChallengeCode.isEmpty()) {
+        connection
+            .prepareStatement("TRUNCATE " + String.join(",", tablesNotChallengeCode) + " CASCADE")
+            .execute();
+      }
 
       connection.commit();
